requirements.txt
# Core Dependencies
requests>=2.31.0
pydantic>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
structlog>=23.0.0
tenacity>=8.2.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
asyncpg>=0.28.0

# Async Support
asyncio>=3.4.3
aiofiles>=23.0.0
uvloop>=0.17.0

# Monitoring and Metrics
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Security
cryptography>=41.0.0
PyJWT>=2.8.0

# Configuration
pyyaml>=6.0
click>=8.1.0

# Utilities
python-dateutil>=2.8.0
validators>=0.20.0
backoff>=2.2.0

# Development (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0
responses>=0.23.0
